+++ src/components/Navbar/Navbar.tsx
@@ -1,12 +1,21 @@
-import { Avatar, Button, Link, Popover, Text } from "@geist-ui/react";
+import {
+	Avatar,
+	Button,
+	Link,
+	Popover,
+	Spacer,
+	Tag,
+	Text,
+} from "@geist-ui/react";
 import { useOnGet } from "@typesaurus/react";
 import React from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { Link as RouterLink } from "react-router-dom";
 import { collection } from "typesaurus";
 
+import _package from "../../../package.json";
 import { ReactComponent as Logo } from "../../assets/logos/white.svg";
-import { auth } from "../../config/firebase";
+import { analytics, auth } from "../../config/firebase";
 import { DiscordIcon, InstagramIcon } from "../../config/icons";
 import { FirestoreUser } from "../../typings/database/User";
 import {
@@ -45,12 +54,25 @@ const Navbar: React.FC<NavbarProperties> = ({
 				</Links>
 				<Actions>
 					<Link
+						onClick={() =>
+							analytics.logEvent("external_link_view", {
+								target: "instagram",
+							})
+						}
 						href="https://www.instagram.com/padova_longboard/"
 						target="_blank"
 					>
 						<InstagramIcon />
 					</Link>
-					<Link href="https://discord.gg/bp7ztg4AEz" target="_blank">
+					<Link
+						onClick={() =>
+							analytics.logEvent("external_link_view", {
+								target: "discord",
+							})
+						}
+						href="https://discord.gg/bp7ztg4AEz"
+						target="_blank"
+					>
 						<DiscordIcon />
 					</Link>
 				</Actions>
@@ -61,9 +83,13 @@ const Navbar: React.FC<NavbarProperties> = ({
 							content={
 								<>
 									<Popover.Item title>
-										<Text
-											span
-										>{`${profile.data.firstName} ${profile.data.lastName}`}</Text>
+										<Text span>{`${
+											profile.data.firstName ? profile.data.firstName : ""
+										} ${
+											profile.data.lastName ? profile.data.lastName : ""
+										}`}</Text>
+										<Spacer style={{ marginLeft: "auto" }} />
+										<Tag>v{_package.version}</Tag>
 									</Popover.Item>
 									<Popover.Item>
 										<Button type="secondary" onClick={onLogoutClick}>