+++ src/components/SocialShareButton/SocialShareButton.tsx
@@ -8,17 +8,19 @@ import { Bullet } from "./styles";
 export interface SocialShareButtonProperties {
 	social: SupportedSocial;
 	target: string;
+	onClick?: (event: React.MouseEvent) => void;
 }
 
 const SocialShareButton: React.FC<SocialShareButtonProperties> = ({
 	social,
 	target,
+	onClick,
 }: SocialShareButtonProperties) => {
 	const icon = SocialIcons[social];
 	const color = SocialColors[social];
 
 	return (
-		<Bullet $color={color} href={target} target="_blank">
+		<Bullet $color={color} href={target} target="_blank" onClick={onClick}>
 			{icon}
 		</Bullet>
 	);