+++ src/index.tsx
@@ -1,9 +1,18 @@
+import * as Sentry from "@sentry/react";
+import { Integrations } from "@sentry/tracing";
 import React from "react";
 import ReactDOM from "react-dom";
 
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
+
+Sentry.init({
+	dsn:
+		"https://a7488499d39c4c7f860304d2a77c8b92@o233916.ingest.sentry.io/5653362",
+	integrations: [new Integrations.BrowserTracing()],
+	tracesSampleRate: 0.25,
+});
 
 ReactDOM.render(
 	<React.StrictMode>
@@ -12,12 +21,6 @@ ReactDOM.render(
 	document.querySelector("#root"),
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://cra.link/PWA
 serviceWorkerRegistration.register();
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();